+++ Platformer/player.cpp
@@ -217,7 +217,7 @@ void Player::update(sf::Time deltaTime, sf::RenderWindow &window, sf::View *view
 	//printf("Total Time: %4.3f\n", (float)total_time);
 
 	// Gravity
-	if (placeFree(x, y + 1, objects)) dy += gravity * mstime;
+	if (placeFree(x, y + 1, objects)) dy += gravity * (float)mstime;
 	else if (dy > 0.0f)
 	{
 		dy = 0.0f;
@@ -240,7 +240,7 @@ void Player::update(sf::Time deltaTime, sf::RenderWindow &window, sf::View *view
 	//if (dy > maxFallSpeed) dy = maxFallSpeed;
 
 	// Update Y
-	for (float i = fabs(dy) * mstime; i > 0; i--)
+	for (float i = fabs(dy) * (float)mstime; i > 0; i--)
 	{
 		float j = copysign(i, dy);
 		if (placeFree(x, y + j, objects))
@@ -253,7 +253,7 @@ void Player::update(sf::Time deltaTime, sf::RenderWindow &window, sf::View *view
 	// Update X
 	if ((dy == 0.0f && !midThrow && !crouching) || jumped)
 	{
-		for (float i = fabs(dx) * mstime; i > 0; i--)
+		for (float i = fabs(dx) * (float)mstime; i > 0; i--)
 		{
 			float j = copysign(i, dx);
 			if (placeFree(x + j, y, objects))
